various bugfixes and optimizations in sorter and outputfile format
git-svn-id: http://biofs04/svn/bioinf/anfo/trunk@1545 93d9a472-daf7-0310-8f21-d7befa554562

+++ anfo-standalone.cc
@@ -509,7 +509,7 @@ int main_( int argc, const char * argv[] )
 		for( int i = 0 ; i != nxthreads ; ++i )
 			pthread_create( indexer_thread+i, 0, run_indexer_thread, &common_data ) ;
 
-	for( int total_count = 0 ; !exit_with && !files.empty() ; files.pop_front() )
+	for( size_t total_count = 0 ; !exit_with && !files.empty() ; files.pop_front() )
 	{
 		// ifstream input_file_stream ;
 		igzstream input_file_stream ;
